refactor: tests

+++ src/Lab1/SpaceTravel/Models/Distance.cs
@@ -1,3 +1,3 @@
 namespace Itmo.ObjectOrientedProgramming.Lab1.SpaceTravel.Models;
 
-public record Distance(int Km);
+public record Distance(double Km);
